tool_cfgable: Moved easy handle cleanup to fix pingpong logout issues
Commmit 1988e601f5 removed the easy handle clean-up from tool_operate,
letting the code that was already present in free_config_fields()
perform the task. Unfortunately, this wasn't the correct place to do
this as it broke protocols, that would perform a logout, as the main
clean-up in tool_main had already been called.